Bug fix and adding hint for other bug
pushed_cursor decreases below zero overflowing and thus becoming very large.
This only occurs when you first debug in embedded mode and generate events
handled by the callbacks on the MCU. This causes afterwards the events not
to be send to the emulator when in proxy mode.

Added also temporary print statements

Added comments for a potential bug related to generating invalid stackValue
type when resolving an event